fix: WebInputHandler 파일 분리
web 빌드 하는 과정에서 readline 에러가 나서 파일을 분리했습니다.
