+++ src/view/InputHandler.js
@@ -1,7 +1,7 @@
 import readLineAsync from "../util/readLineAsync.js";
 import OutputView from "./outputView.js";
 
-export const InputHandler = async ({ inputMessage, parser, validator }) => {
+const InputHandler = async ({ inputMessage, parser, validator }) => {
   try {
     const input = await readLineAsync(inputMessage);
     const parsedInput = parser ? parser(input) : input;
@@ -13,19 +13,4 @@ export const InputHandler = async ({ inputMessage, parser, validator }) => {
   }
 };
 
-export const WebInputHandler = ({ elementIds, parser, validator }) => {
-  if (typeof elementIds === "string") {
-    const input = document.getElementById(elementIds).value;
-    const parsedInput = parser ? parser(input) : input;
-    validator(parsedInput);
-    return parsedInput;
-  }
-  if (Array.isArray(elementIds)) {
-    const inputs = elementIds.map((id) => document.getElementById(id).value);
-    const parsedInputs = inputs.map((input) =>
-      parser ? parser(input) : input
-    );
-    validator(parsedInputs);
-    return parsedInputs;
-  }
-};
+export default InputHandler;
